Moved common CLI decorators to separate module + test update
